+++ src/jk_logging/__init__.py
@@ -1,8 +1,8 @@
-
+﻿
 
 
 __author__ = "Jürgen Knauth"
-__version__ = "0.2022.3.27"
+__version__ = "0.2022.3.31"